% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytofkitNormalization.R
\name{CytofkitNormalization}
\alias{CytofkitNormalization}
\title{CytofkitNormalization}
\description{
A template for building documented, tested R6 classes
}
\examples{
  rt <- R6Template$new(id="abc")

## ------------------------------------------------
## Method `CytofkitNormalization$createTableForViolinPlot`
## ------------------------------------------------

tbl.violin <- createTableForViolinPlot(1:20, "H3")
ggplot(tbl.violin, aes(x=name, y=value, fill=name)) +
    geom_violin() +
    coord_cartesian(ylim=c(-5,5)) +
    theme(axis.text = element_text(size = 14)) +
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CytofkitNormalization-new}{\code{CytofkitNormalization$new()}}
\item \href{#method-CytofkitNormalization-getMatrix}{\code{CytofkitNormalization$getMatrix()}}
\item \href{#method-CytofkitNormalization-getTsne}{\code{CytofkitNormalization$getTsne()}}
\item \href{#method-CytofkitNormalization-getClusters}{\code{CytofkitNormalization$getClusters()}}
\item \href{#method-CytofkitNormalization-createSimpleMarkerNames}{\code{CytofkitNormalization$createSimpleMarkerNames()}}
\item \href{#method-CytofkitNormalization-getMarkers}{\code{CytofkitNormalization$getMarkers()}}
\item \href{#method-CytofkitNormalization-getCluster}{\code{CytofkitNormalization$getCluster()}}
\item \href{#method-CytofkitNormalization-normalizeMarker}{\code{CytofkitNormalization$normalizeMarker()}}
\item \href{#method-CytofkitNormalization-createTableForViolinPlot}{\code{CytofkitNormalization$createTableForViolinPlot()}}
\item \href{#method-CytofkitNormalization-calculateColorBoundaries}{\code{CytofkitNormalization$calculateColorBoundaries()}}
\item \href{#method-CytofkitNormalization-createTableForTsnePlot}{\code{CytofkitNormalization$createTableForTsnePlot()}}
\item \href{#method-CytofkitNormalization-clone}{\code{CytofkitNormalization$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-new"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$new(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{character, the full path to cytofkit results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of CytofkitNormalization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getMatrix}{}}}
\subsection{Method \code{getMatrix()}}{
accessor for the object's filename
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getMatrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the current value of the filename
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getTsne"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getTsne}{}}}
\subsection{Method \code{getTsne()}}{
accessor for the object's tbl.tsne
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getTsne()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getClusters"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getClusters}{}}}
\subsection{Method \code{getClusters()}}{
accessor for the object's cluster list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getClusters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-createSimpleMarkerNames"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-createSimpleMarkerNames}{}}}
\subsection{Method \code{createSimpleMarkerNames()}}{
accessor for markers used in the cytof
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$createSimpleMarkerNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getMarkers"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getMarkers}{}}}
\subsection{Method \code{getMarkers()}}{
accessor for named list, e.g. list("H3"="Yb176Di<176Yb_H3>")
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getMarkers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getCluster"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getCluster}{}}}
\subsection{Method \code{getCluster()}}{
accessor for named list, e.g. list("H3"="Yb176Di<176Yb_H3>")
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getCluster(clusterNumber)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusterNumber}}{numeric - must be in range}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-normalizeMarker"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-normalizeMarker}{}}}
\subsection{Method \code{normalizeMarker()}}{
calculate and append a normalized vector column for the specified
  histone marker, regressed against the base ("total") H3 and/or H4 vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$normalizeMarker(target.marker, reference.markers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.marker}}{character, a column name from the matrix, e.g., "Sm154Di<154Sm_H3K27me2>"}

\item{\code{reference.markers}}{character, the marker names for reference ("total") H3 and/or H4}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character the newly created column name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-createTableForViolinPlot"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-createTableForViolinPlot}{}}}
\subsection{Method \code{createTableForViolinPlot()}}{
create a ggplot2 violin plot data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$createTableForViolinPlot(clusters, marker, matrix = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusters}}{numeric a vector of cluster numbers}

\item{\code{marker}}{character, the marker name}

\item{\code{matrix}}{matrix, a subset of the full matrix, used only if you wish to make violins
out of different classes of cells in each cluster see unitTests}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame ready for call to ggplot
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{tbl.violin <- createTableForViolinPlot(1:20, "H3")
ggplot(tbl.violin, aes(x=name, y=value, fill=name)) +
    geom_violin() +
    coord_cartesian(ylim=c(-5,5)) +
    theme(axis.text = element_text(size = 14)) +
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-calculateColorBoundaries"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-calculateColorBoundaries}{}}}
\subsection{Method \code{calculateColorBoundaries()}}{
calculates color boundaries for a distribution of numeric values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$calculateColorBoundaries(vector, colors, mode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vector}}{numeric}

\item{\code{colors}}{character, a vector of color names, e.g. "#5E4FA2" "#3288BD" ...}

\item{\code{mode}}{character either "quantile" or "interval"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame with min, max, color columns
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-createTableForTsnePlot"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-createTableForTsnePlot}{}}}
\subsection{Method \code{createTableForTsnePlot()}}{
create a tsne (x,y,color) data.frame with specified color mapping,
   for the specified marker
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$createTableForTsnePlot(marker, colors, matrix.sub = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{marker}}{character, the (short) marker name}

\item{\code{colors}}{character, a vector of color names, e.g. "#5E4FA2" "#3288BD" ...}

\item{\code{matrix.sub}}{matrix, an optional subset of the full matrix,}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list data.frame (for tsne plotting) and color quantiles (for plot legend)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
