% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytofkitNormalization.R
\name{CytofkitNormalization}
\alias{CytofkitNormalization}
\title{CytofkitNormalization}
\description{
A template for building documented, tested R6 classes
}
\examples{
  rt <- R6Template$new(id="abc")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CytofkitNormalization-new}{\code{CytofkitNormalization$new()}}
\item \href{#method-CytofkitNormalization-getMatrix}{\code{CytofkitNormalization$getMatrix()}}
\item \href{#method-CytofkitNormalization-getTsne}{\code{CytofkitNormalization$getTsne()}}
\item \href{#method-CytofkitNormalization-getClusters}{\code{CytofkitNormalization$getClusters()}}
\item \href{#method-CytofkitNormalization-clone}{\code{CytofkitNormalization$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-new"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$new(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{character, the full path to cytofkit results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of CytofkitNormalization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getMatrix}{}}}
\subsection{Method \code{getMatrix()}}{
accessor for the object's filename
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getMatrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the current value of the filename
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getTsne"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getTsne}{}}}
\subsection{Method \code{getTsne()}}{
accessor for the object's tbl.tsne
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getTsne()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getClusters"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getClusters}{}}}
\subsection{Method \code{getClusters()}}{
accessor for the object's cluster list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getClusters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
