% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytofkitNormalization.R
\name{CytofkitNormalization}
\alias{CytofkitNormalization}
\title{CytofkitNormalization}
\description{
A template for building documented, tested R6 classes
}
\examples{
  rt <- R6Template$new(id="abc")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CytofkitNormalization-new}{\code{CytofkitNormalization$new()}}
\item \href{#method-CytofkitNormalization-getMatrix}{\code{CytofkitNormalization$getMatrix()}}
\item \href{#method-CytofkitNormalization-getTsne}{\code{CytofkitNormalization$getTsne()}}
\item \href{#method-CytofkitNormalization-getClusters}{\code{CytofkitNormalization$getClusters()}}
\item \href{#method-CytofkitNormalization-createSimpleMarkerNames}{\code{CytofkitNormalization$createSimpleMarkerNames()}}
\item \href{#method-CytofkitNormalization-getMarkers}{\code{CytofkitNormalization$getMarkers()}}
\item \href{#method-CytofkitNormalization-normalizeMarker}{\code{CytofkitNormalization$normalizeMarker()}}
\item \href{#method-CytofkitNormalization-clone}{\code{CytofkitNormalization$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-new"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$new(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{character, the full path to cytofkit results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of CytofkitNormalization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getMatrix}{}}}
\subsection{Method \code{getMatrix()}}{
accessor for the object's filename
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getMatrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the current value of the filename
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getTsne"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getTsne}{}}}
\subsection{Method \code{getTsne()}}{
accessor for the object's tbl.tsne
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getTsne()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getClusters"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getClusters}{}}}
\subsection{Method \code{getClusters()}}{
accessor for the object's cluster list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getClusters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-createSimpleMarkerNames"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-createSimpleMarkerNames}{}}}
\subsection{Method \code{createSimpleMarkerNames()}}{
accessor for markers used in the cytof
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$createSimpleMarkerNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-getMarkers"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-getMarkers}{}}}
\subsection{Method \code{getMarkers()}}{
accessor for named list, e.g. list("H3"="Yb176Di<176Yb_H3>")
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$getMarkers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-normalizeMarker"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-normalizeMarker}{}}}
\subsection{Method \code{normalizeMarker()}}{
calculate and append a normalized vector column for the specified
  histone marker, regressed against the base ("total") H3 and/or H4 vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$normalizeMarker(target.marker, reference.markers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.marker}}{character, a column name from the matrix, e.g., "Sm154Di<154Sm_H3K27me2>"}

\item{\code{reference.markers}}{character, the marker names for reference ("total") H3 and/or H4}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character the newly created column name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CytofkitNormalization-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CytofkitNormalization-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CytofkitNormalization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
